[parity][opt] Use (almost) fixed-size frames
Instead of pushing arguments on to the frame and then poping
them after the call reserve a space for outgoing arguments
above spill slots and put arguments into it.

This allows to remove arguments poping after the call and accounts
for significant drop in code size (>240K on Flutter Gallery). There is
a slight increase in RX size due to naive approach to stack maps
which can be mitigated later.