[vm] Support breakpoint updates in reloadSources
Introduce breakpointsUpdates parameter for reloadSources service
method.

If this parameter is provided, then the given updates (breakpoint
removals and additions) will be atomically applied after the successful
reload if reload changed any libraries (i.e. the reload was not a no-op)
and before isolates in the group start running again.

Bump protocol version to 4.21.

TEST=pkg/vm_service/reload_sources_test

Change-Id: I526db458de62bacddd9d8f7e333c1fccf77a250c